Create a separate context for pipelines
